[fix] 재발급, 로그아웃 컨트롤러 변경
