+++ src/main/java/modernfarmer/server/farmususer/user/controller/AuthController.java
@@ -35,32 +35,32 @@ public class AuthController {
 
         return reissueTokenResponseDto;
     }
-
-    @DeleteMapping("/logout")
-    public ResponseDto logout(HttpServletRequest request)  {
-
-
-        String userId = jwtTokenProvider.getUserId(request);
-
-        ResponseDto logoutResponseDto = authService.logout(Long.valueOf(userId));
-
-        LOGGER.info("로그아웃 완료");
-
-        return logoutResponseDto;
-    }
-
-    @GetMapping(value = "/reissue-token")
-    public TokenResponseDto reissueToken(HttpServletRequest request)  {
-
-        String userId = jwtTokenProvider.getUserId(request);
-        String refreshToken = jwtTokenProvider.resolveToken(request);
-
-        TokenResponseDto reissueTokenResponseDto = authService.reissueToken(refreshToken, Long.valueOf(userId));
-
-        LOGGER.info("토큰 재발급 완료");
-
-        return reissueTokenResponseDto;
-    }
+//
+//    @DeleteMapping("/logout")
+//    public ResponseDto logout(HttpServletRequest request)  {
+//
+//
+//        String userId = jwtTokenProvider.getUserId(request);
+//
+//        ResponseDto logoutResponseDto = authService.logout(Long.valueOf(userId));
+//
+//        LOGGER.info("로그아웃 완료");
+//
+//        return logoutResponseDto;
+//    }
+//
+//    @GetMapping(value = "/reissue-token")
+//    public TokenResponseDto reissueToken(HttpServletRequest request)  {
+//
+//        String userId = jwtTokenProvider.getUserId(request);
+//        String refreshToken = jwtTokenProvider.resolveToken(request);
+//
+//        TokenResponseDto reissueTokenResponseDto = authService.reissueToken(refreshToken, Long.valueOf(userId));
+//
+//        LOGGER.info("토큰 재발급 완료");
+//
+//        return reissueTokenResponseDto;
+//    }
 
     @PostMapping(value = "/google-login")
     public TokenResponseDto googleLogin(HttpServletRequest request)  {
